requirements.txt
requests>=2.14.2
GitPython
appdirs
